cloud_function-bigquery_to_cloudstorage/requirements.txt
# Function dependencies, for example:
# package>=version


google-cloud-bigquery
